refactor: create seperate imagesController
