+++ portfolio/src/app/api/gallery/collectionsController.ts
@@ -52,38 +52,7 @@ export async function createEmptyCollection(title: string) {
   return newCollection;
 }
 
-export async function createImage(fileName: string, filePath: string) {
-  const newImage = await prisma.image.create({
-    data: {
-      fileName: fileName,
-      filePath: filePath,
-    },
-  });
-
-  return newImage;
-}
-
-export async function addImageToCollection(
-  collectionId: number,
-  imageId: number
-) {
-  const collection = await prisma.collection.update({
-    where: { id: collectionId },
-    data: {
-      images: {
-        connect: {
-          id: imageId,
-        },
-      },
-    },
-  });
-
-  await updateCollectionPreview(collectionId);
-
-  return collection;
-}
-
-async function updateCollectionPreview(collectionId: number) {
+export async function updateCollectionPreview(collectionId: number) {
   const images = await prisma.collection.findFirst({
     where: { id: collectionId },
     include: {
